+++ examples/custom_convert.py
@@ -5,7 +5,6 @@ from pathlib import Path
 from typing import Iterable
 
 from docling.backend.docling_parse_backend import DoclingParseDocumentBackend
-from docling.backend.pypdfium2_backend import PyPdfiumDocumentBackend
 from docling.datamodel.base_models import ConversionStatus, PdfPipelineOptions
 from docling.datamodel.document import ConversionResult, DocumentConversionInput
 from docling.pdf_document_converter import PdfDocumentConverter
@@ -28,19 +27,23 @@ def export_documents(
             doc_filename = conv_res.input.file.stem
 
             # Export Deep Search document JSON format:
-            with (output_dir / f"{doc_filename}.json").open("w") as fp:
+            with (output_dir / f"{doc_filename}.json").open(
+                "w", encoding="utf-8"
+            ) as fp:
                 fp.write(json.dumps(conv_res.render_as_dict()))
 
             # Export Text format:
-            with (output_dir / f"{doc_filename}.txt").open("w") as fp:
+            with (output_dir / f"{doc_filename}.txt").open("w", encoding="utf-8") as fp:
                 fp.write(conv_res.render_as_text())
 
             # Export Markdown format:
-            with (output_dir / f"{doc_filename}.md").open("w") as fp:
+            with (output_dir / f"{doc_filename}.md").open("w", encoding="utf-8") as fp:
                 fp.write(conv_res.render_as_markdown())
 
             # Export Document Tags format:
-            with (output_dir / f"{doc_filename}.doctags").open("w") as fp:
+            with (output_dir / f"{doc_filename}.doctags").open(
+                "w", encoding="utf-8"
+            ) as fp:
                 fp.write(conv_res.render_as_doctags())
 
         else:
@@ -82,7 +85,7 @@ def main():
     # PyPdfium with OCR
     # -----------------
     # pipeline_options = PipelineOptions()
-    # pipeline_options.do_ocr=False
+    # pipeline_options.do_ocr=True
     # pipeline_options.do_table_structure=True
     # pipeline_options.table_structure_options.do_cell_matching = True